+++ src/pages/Profile/bookmark/BookmarkPage.tsx
@@ -4,8 +4,8 @@ import * as Factory from 'factory.ts';
 // import EmptyBookmarkPage from './emptyBookmark/EmptyBookmarkPage';
 import './BookmarkPage.scss';
 
-import Study from './study/Study';
-import Article from './article/Article';
+import Study from '../../../app/profile/bookmark/study/Study';
+import Article from '../../../app/profile/bookmark/article/Article';
 
 const toFormattedCount = (count: number) => {
 	if (count < 10) return `0${count}`;
